requirements.txt
openai
langchain
langgraph
tavily-python
faiss-cpu
streamlit  # Optional for UI
jupyter    # Optional for testing